+++ run_tests.py
@@ -10,6 +10,7 @@ if __name__ == '__main__':
         '--cov-report', 'term',
         '--xdoctest',
         '--cov=' + package_name,
+        package_name, 'tests'
     ]
     pytest_args = pytest_args + sys.argv[1:]
     sys.exit(pytest.main(pytest_args))