+++ ubelt/util_arg.py
@@ -12,9 +12,17 @@ The :func:`argval` function returns the value of a ``--key=value`` style CLI
 argument.
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
-import six
 import sys
 from ubelt import util_const
+
+
+PY2 = sys.version_info[0] == 2
+
+if PY2:
+    import six
+    string_types = six.string_types
+else:
+    string_types = (str,)
 
 
 def argval(key, default=util_const.NoParam, argv=None):
@@ -60,7 +68,7 @@ def argval(key, default=util_const.NoParam, argv=None):
     if argv is None:  # nocover
         argv = sys.argv
 
-    keys = [key] if isinstance(key, six.string_types) else key
+    keys = [key] if isinstance(key, string_types) else key
     n_max = len(argv) - 1
     for argx, item in enumerate(argv):
         for key_ in keys:
@@ -101,6 +109,6 @@ def argflag(key, argv=None):
     """
     if argv is None:  # nocover
         argv = sys.argv
-    keys = [key] if isinstance(key, six.string_types) else key
+    keys = [key] if isinstance(key, string_types) else key
     flag = any(k in argv for k in keys)
     return flag