+++ ubelt/util_dict.py
@@ -28,14 +28,24 @@ The :func:`ddict` and :func:`odict` functions are alias for the commonly used
 
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
-import six
+import sys
 import operator as op
 import itertools as it
 from collections import OrderedDict
 from collections import defaultdict
-from six.moves import zip
 from ubelt import util_const
 from ubelt import util_list
+
+
+PY2 = sys.version_info[0] == 2
+
+if PY2:
+    import six
+    from six.moves import zip
+    iteritems = six.iteritems
+else:
+    def iteritems(d, **kw):
+        return d.items(**kw)
 
 
 # Expose for convenience
@@ -522,7 +532,7 @@ def map_vals(func, dict_):
     """
     if not hasattr(func, '__call__'):
         func = func.__getitem__
-    keyval_list = [(key, func(val)) for key, val in six.iteritems(dict_)]
+    keyval_list = [(key, func(val)) for key, val in iteritems(dict_)]
     dictclass = OrderedDict if isinstance(dict_, OrderedDict) else dict
     newdict = dictclass(keyval_list)
     return newdict
@@ -559,7 +569,7 @@ def map_keys(func, dict_):
     """
     if not hasattr(func, '__call__'):
         func = func.__getitem__
-    keyval_list = [(func(key), val) for key, val in six.iteritems(dict_)]
+    keyval_list = [(func(key), val) for key, val in iteritems(dict_)]
     dictclass = OrderedDict if isinstance(dict_, OrderedDict) else dict
     newdict = dictclass(keyval_list)
     if len(newdict) != len(dict_):